+++ database/redisearch_alter.go
@@ -25,7 +25,7 @@ func execFTAlter(db *DB, args [][]byte) redis.Reply {
 	engine, ok := searchEngines[indexName]
 	searchEnginesMu.RUnlock()
 	if !ok || engine == nil {
-		return protocol.MakeErrReply(fmt.Sprintf("ERR Index '%s' does not exist", string(args[0])))
+		return protocol.MakeErrReply(fmt.Sprintf("SEARCH_INDEX_NOT_FOUND Index not found: %s", string(args[0])))
 	}
 
 	fields, errReply := parseFTSchemaFields(args[3:])
@@ -75,7 +75,7 @@ func execFTExplain(db *DB, args [][]byte) redis.Reply {
 	_, ok := searchEngines[indexName]
 	searchEnginesMu.RUnlock()
 	if !ok {
-		return protocol.MakeErrReply(fmt.Sprintf("ERR Index '%s' does not exist", string(args[0])))
+		return protocol.MakeErrReply(fmt.Sprintf("SEARCH_INDEX_NOT_FOUND Index not found: %s", string(args[0])))
 	}
 
 	return protocol.MakeBulkReply([]byte(formatFTExplainPlan(query)))